Build macOS wheels for multiple Python versions
Rework macOS wheel build scripts to use Python-version-specific
permanent locations for all operations, with temporary symlinks used for
operations shared with the Linux wheel builds that expect non-versioned
locations. Also rework builder to support a set of Python targets and to
be able to build multiple wheels in a manner similar to how things work
for Linux.

This changes the macOS wheel build to produce wheels for (as of writing)
both Python 3.11 and 3.12.

+++ tools/wheel/macos/provision-test-python.sh
@@ -6,10 +6,17 @@
 
 set -eu -o pipefail
 
+if [[ $# -lt 1 ]]; then
+    echo "Usage: $0 <python-version>" >&2
+    exit 1
+fi
+
 # Clean up from old tests.
 rm -rf /opt/drake-wheel-test
 
 # Prepare test environment.
+mkdir /opt/drake-wheel-test
+
 # NOTE: Xcode ships python3, make sure to use the one from brew.
-$(brew --prefix python@3.11)/bin/python3.11 \
-    -m venv /opt/drake-wheel-test/python
+$(brew --prefix python@$1)/bin/python$1 \
+    -m venv /opt/drake-wheel-test/python$1
